docs: Split snippets into their own files
